+++ protoc-gen-kroto-plus/src/main/kotlin/com/github/marcoferrer/krotoplus/generators/GrpcStubExtsGenerator.kt
@@ -18,6 +18,9 @@ package com.github.marcoferrer.krotoplus.generators
 
 import com.github.marcoferrer.krotoplus.config.GrpcStubExtsGenOptions
 import com.github.marcoferrer.krotoplus.generators.Generator.Companion.AutoGenerationDisclaimer
+import com.github.marcoferrer.krotoplus.generators.builders.ServerStreamingStubExtsBuilder
+import com.github.marcoferrer.krotoplus.generators.builders.UnaryStubExtsBuilder
+import com.github.marcoferrer.krotoplus.generators.builders.addResponseObserverParamerter
 import com.github.marcoferrer.krotoplus.proto.ProtoMethod
 import com.github.marcoferrer.krotoplus.proto.ProtoService
 import com.github.marcoferrer.krotoplus.proto.Schema
@@ -137,328 +140,3 @@ object GrpcStubExtsGenerator : Generator {
     }
 }
 
-class ServerStreamingStubExtsBuilder(val context: GeneratorContext){
-
-    fun buildStubExts(method: ProtoMethod, options: GrpcStubExtsGenOptions): List<FunSpec> {
-        val funSpecs = mutableListOf<FunSpec>()
-
-        // Add method signature exts
-        if(method.methodSignatureFields.isNotEmpty()){
-            funSpecs += buildAsyncMethodSigExt(method)
-            funSpecs += buildBlockingMethodSigExt(method)
-        }
-
-        // Add lambda builder exts
-        funSpecs += buildAsyncLambdaBuilderExt(method)
-        funSpecs += buildBlockingLambdaBuilderExt(method)
-
-        // Add default arg exts
-        funSpecs += buildAsyncDefaultArgExt(method)
-        funSpecs += buildBlockingDefaultArgExt(method)
-
-        if(options.supportCoroutines)
-            addCoroutineStubExts(funSpecs, method)
-
-        return funSpecs
-    }
-
-    private fun addCoroutineStubExts(funSpecs: MutableList<FunSpec>, method: ProtoMethod){
-        if(method.methodSignatureFields.isNotEmpty()) {
-            funSpecs += buildCoroutineMethodSigExt(method)
-        }
-
-        funSpecs += buildCoroutineExt(method)
-        funSpecs += buildCoroutineLambdaBuilderExt(method)
-    }
-
-    // Coroutine Extension
-
-    private fun buildCoroutineExt(protoMethod: ProtoMethod): FunSpec = with(protoMethod){
-        FunSpec.builder(functionName)
-            .returns(CommonClassNames.receiveChannel.parameterizedBy(responseClassName))
-            .receiver(protoService.asyncStubClassName)
-            .addParameter(requestClassName.requestParamSpec)
-            .addStatement(
-                "return %T(request, %T.%N())",
-                CommonClassNames.ClientCalls.clientCallServerStreaming,
-                protoService.enclosingServiceClassName,
-                methodDefinitionGetterName
-            )
-            .build()
-    }
-
-    // Method Signature Extensions
-
-    private fun buildAsyncMethodSigExt(protoMethod: ProtoMethod): FunSpec = with(protoMethod){
-        FunSpec.builder(functionName)
-            .receiver(protoService.asyncStubClassName)
-            .returns(UNIT)
-            .addMethodSignatureParamerter(methodSignatureFields, context.schema)
-            .addResponseObserverParamerter(responseClassName)
-            .addCode(requestClassName.requestValueMethodSigCodeBlock(methodSignatureFields))
-            .addStatement("%N(request, responseObserver)",functionName)
-            .build()
-    }
-
-    private fun buildCoroutineMethodSigExt(protoMethod: ProtoMethod): FunSpec = with(protoMethod){
-        FunSpec.builder(functionName)
-            .receiver(protoService.asyncStubClassName)
-            .addMethodSignatureParamerter(methodSignatureFields, context.schema)
-            .returns(CommonClassNames.receiveChannel.parameterizedBy(responseClassName))
-            .addCode(requestClassName.requestValueMethodSigCodeBlock(methodSignatureFields))
-            .addStatement("return %N(request)",functionName)
-            .build()
-    }
-
-    private fun buildBlockingMethodSigExt(protoMethod: ProtoMethod): FunSpec = with(protoMethod){
-        FunSpec.builder(functionName)
-            .receiver(protoService.blockingStubClassName)
-            .addMethodSignatureParamerter(methodSignatureFields, context.schema)
-            .returns(Iterator::class.asClassName().parameterizedBy(responseClassName))
-            .addCode(requestClassName.requestValueMethodSigCodeBlock(methodSignatureFields))
-            .addStatement("return %N(request)",functionName)
-            .build()
-    }
-
-    // Lambda Builder Extensions
-
-    private fun buildAsyncLambdaBuilderExt(protoMethod: ProtoMethod): FunSpec = with(protoMethod){
-        FunSpec.builder(functionName)
-            .addModifiers(KModifier.INLINE)
-            .receiver(protoService.asyncStubClassName)
-            .addResponseObserverParamerter(responseClassName)
-            .addParameter("block", requestClassName.builderLambdaTypeName)
-            .returns(UNIT)
-            .addCode(requestClassName.requestValueBuilderCodeBlock)
-            .addStatement("%N(request, responseObserver)", functionName)
-            .build()
-    }
-
-    private fun buildCoroutineLambdaBuilderExt(protoMethod: ProtoMethod): FunSpec = with(protoMethod){
-        FunSpec.builder(functionName)
-            .receiver(protoService.asyncStubClassName)
-            .returns(CommonClassNames.receiveChannel.parameterizedBy(responseClassName))
-            .addModifiers(KModifier.INLINE)
-            .addParameter("block", requestClassName.builderLambdaTypeName)
-            .addCode(requestClassName.requestValueBuilderCodeBlock)
-            .addStatement("return %N(request)", functionName)
-            .build()
-    }
-
-    private fun buildBlockingLambdaBuilderExt(protoMethod: ProtoMethod): FunSpec = with(protoMethod){
-        FunSpec.builder(functionName)
-            .addModifiers(KModifier.INLINE)
-            .receiver(protoService.blockingStubClassName)
-            .addParameter("block", requestClassName.builderLambdaTypeName)
-            .returns(Iterator::class.asClassName().parameterizedBy(responseClassName))
-            .addCode(requestClassName.requestValueBuilderCodeBlock)
-            .addStatement("return %N(request)", functionName)
-            .build()
-    }
-
-    // Default Argument Extensions
-
-    private fun buildAsyncDefaultArgExt(protoMethod: ProtoMethod): FunSpec = with(protoMethod){
-        FunSpec.builder(functionName)
-            .receiver(protoService.asyncStubClassName)
-            .addResponseObserverParamerter(responseClassName)
-            .returns(UNIT)
-            .addStatement("%N(%T.getDefaultInstance(),responseObserver)", functionName, requestClassName)
-            .build()
-    }
-
-    private fun buildBlockingDefaultArgExt(protoMethod: ProtoMethod): FunSpec = with(protoMethod){
-        FunSpec.builder(functionName)
-            .receiver(protoService.blockingStubClassName)
-            .returns(Iterator::class.asClassName().parameterizedBy(responseClassName))
-            .addStatement("return %N(%T.getDefaultInstance())", functionName, requestClassName)
-            .build()
-    }
-
-}
-
-class UnaryStubExtsBuilder(val context: GeneratorContext){
-
-    fun buildStubExts(method: ProtoMethod, options: GrpcStubExtsGenOptions): List<FunSpec> {
-        val funSpecs = mutableListOf<FunSpec>()
-
-        // Add method signature exts
-        if(method.methodSignatureFields.isNotEmpty()){
-            funSpecs += buildAsyncMethodSigExt(method)
-            funSpecs += buildFutureMethodSigExt(method)
-            funSpecs += buildBlockingMethodSigExt(method)
-        }
-
-        // Add lambda builder exts
-        funSpecs += buildAsyncLambdaBuilderExt(method)
-        funSpecs += buildFutureLambdaBuilderExt(method)
-        funSpecs += buildBlockingLambdaBuilderExt(method)
-
-        // Add default arg exts
-        funSpecs += buildAsyncDefaultArgExt(method)
-        funSpecs += buildFutureDefaultArgExt(method)
-        funSpecs += buildBlockingDefaultArgExt(method)
-
-        if(options.supportCoroutines)
-            addCoroutineStubExts(funSpecs, method)
-
-        return funSpecs
-    }
-
-    private fun addCoroutineStubExts(funSpecs: MutableList<FunSpec>, method: ProtoMethod){
-        if(method.methodSignatureFields.isNotEmpty()) {
-            funSpecs += buildCoroutineMethodSigExt(method)
-        }
-
-        funSpecs += buildCoroutineExt(method)
-        funSpecs += buildCoroutineLambdaBuilderExt(method)
-    }
-
-    // Coroutine Extension
-
-    private fun buildCoroutineExt(protoMethod: ProtoMethod): FunSpec = with(protoMethod){
-        FunSpec.builder(functionName)
-            .addModifiers(KModifier.SUSPEND)
-            .receiver(protoService.asyncStubClassName)
-            .addParameter(requestClassName.requestParamSpec)
-            .returns(responseClassName)
-            .addStatement(
-                "return %T(request, %T.%N())",
-                CommonClassNames.ClientCalls.clientCallUnary,
-                protoService.enclosingServiceClassName,
-                methodDefinitionGetterName
-            )
-            .build()
-    }
-
-    // Method Signature Extensions
-
-    private fun buildAsyncMethodSigExt(protoMethod: ProtoMethod): FunSpec = with(protoMethod){
-        FunSpec.builder(functionName)
-            .receiver(protoService.asyncStubClassName)
-            .addMethodSignatureParamerter(methodSignatureFields, context.schema)
-            .addResponseObserverParamerter(responseClassName)
-            .returns(UNIT)
-            .addCode(requestClassName.requestValueMethodSigCodeBlock(methodSignatureFields))
-            .addStatement("%N(request, responseObserver)",functionName)
-            .build()
-    }
-
-    private fun buildCoroutineMethodSigExt(protoMethod: ProtoMethod): FunSpec = with(protoMethod){
-        FunSpec.builder(functionName)
-            .addModifiers(KModifier.SUSPEND)
-            .receiver(protoService.asyncStubClassName)
-            .returns(responseClassName)
-            .addMethodSignatureParamerter(methodSignatureFields, context.schema)
-            .addCode(requestClassName.requestValueMethodSigCodeBlock(methodSignatureFields))
-            .addStatement("return %N(request)",functionName)
-            .build()
-    }
-
-    private fun buildFutureMethodSigExt(protoMethod: ProtoMethod): FunSpec = with(protoMethod){
-        FunSpec.builder(functionName)
-            .receiver(protoService.futureStubClassName)
-            .addMethodSignatureParamerter(methodSignatureFields, context.schema)
-            .returns(CommonClassNames.listenableFuture.parameterizedBy(responseClassName))
-            .addCode(requestClassName.requestValueMethodSigCodeBlock(methodSignatureFields))
-            .addStatement("return %N(request)",functionName)
-            .build()
-    }
-
-    private fun buildBlockingMethodSigExt(protoMethod: ProtoMethod): FunSpec = with(protoMethod){
-        FunSpec.builder(functionName)
-            .receiver(protoService.blockingStubClassName)
-            .returns(responseClassName)
-            .addMethodSignatureParamerter(methodSignatureFields, context.schema)
-            .addCode(requestClassName.requestValueMethodSigCodeBlock(methodSignatureFields))
-            .addStatement("return %N(request)",functionName)
-            .build()
-    }
-
-    // Lambda Builder Extensions
-
-    private fun buildAsyncLambdaBuilderExt(protoMethod: ProtoMethod): FunSpec = with(protoMethod){
-        FunSpec.builder(functionName)
-            .receiver(protoService.asyncStubClassName)
-            .addModifiers(KModifier.INLINE)
-            .addResponseObserverParamerter(responseClassName)
-            .addParameter("block", requestClassName.builderLambdaTypeName)
-            .returns(UNIT)
-            .addCode(requestClassName.requestValueBuilderCodeBlock)
-            .addStatement("%N(request, responseObserver)", functionName)
-            .build()
-    }
-
-    private fun buildCoroutineLambdaBuilderExt(protoMethod: ProtoMethod): FunSpec = with(protoMethod){
-        FunSpec.builder(functionName)
-            .receiver(protoService.asyncStubClassName)
-            .addModifiers(KModifier.INLINE, KModifier.SUSPEND)
-            .addParameter("block", requestClassName.builderLambdaTypeName)
-            .returns(responseClassName)
-            .addCode(requestClassName.requestValueBuilderCodeBlock)
-            .addStatement("return %N(request)", functionName)
-            .build()
-    }
-
-    private fun buildFutureLambdaBuilderExt(protoMethod: ProtoMethod): FunSpec = with(protoMethod){
-        FunSpec.builder(functionName)
-            .addModifiers(KModifier.INLINE)
-            .receiver(protoService.futureStubClassName)
-            .addParameter("block", requestClassName.builderLambdaTypeName)
-            .returns(CommonClassNames.listenableFuture.parameterizedBy(responseClassName))
-            .addCode(requestClassName.requestValueBuilderCodeBlock)
-            .addStatement("return %N(request)", functionName)
-            .build()
-    }
-
-    private fun buildBlockingLambdaBuilderExt(protoMethod: ProtoMethod): FunSpec = with(protoMethod){
-        FunSpec.builder(functionName)
-            .addModifiers(KModifier.INLINE)
-            .addCode(requestClassName.requestValueBuilderCodeBlock)
-            .addParameter("block", requestClassName.builderLambdaTypeName)
-            .addStatement("return %N(request)", functionName)
-            .receiver(protoService.blockingStubClassName)
-            .returns(responseClassName)
-            .build()
-    }
-
-    // Default Argument Extensions
-
-    private fun buildAsyncDefaultArgExt(protoMethod: ProtoMethod): FunSpec = with(protoMethod){
-        FunSpec.builder(functionName)
-            .receiver(protoService.asyncStubClassName)
-            .addResponseObserverParamerter(responseClassName)
-            .addStatement("%N(%T.getDefaultInstance(),responseObserver)", functionName, requestClassName)
-            .returns(UNIT)
-            .build()
-    }
-
-    private fun buildFutureDefaultArgExt(protoMethod: ProtoMethod): FunSpec = with(protoMethod){
-        FunSpec.builder(functionName)
-            .addStatement("return %N(%T.getDefaultInstance())", functionName, requestClassName)
-            .receiver(protoService.futureStubClassName)
-            .returns(CommonClassNames.listenableFuture.parameterizedBy(responseClassName))
-            .build()
-    }
-
-    private fun buildBlockingDefaultArgExt(protoMethod: ProtoMethod): FunSpec = with(protoMethod){
-        FunSpec.builder(functionName)
-            .addStatement("return %N(%T.getDefaultInstance())", functionName, requestClassName)
-            .receiver(protoService.blockingStubClassName)
-            .returns(responseClassName)
-            .build()
-    }
-}
-
-private fun FunSpec.Builder.addResponseObserverParamerter(responseClassName: ClassName): FunSpec.Builder = apply {
-    addParameter("responseObserver", CommonClassNames.streamObserver.parameterizedBy(responseClassName))
-}
-
-private fun FunSpec.Builder.addMethodSignatureParamerter(
-    methodSignatureFields: List<DescriptorProtos.FieldDescriptorProto>,
-    schema: Schema
-): FunSpec.Builder = apply {
-    addForEach(methodSignatureFields){
-        addParameter(it.name.toUpperCamelCase().decapitalize(), it.getFieldClassName(schema))
-    }
-}
